Introduce NMIs with Ibex's MStack registers
This contains some CHERI specific code, so this commit is not really
applicable to pure Ibex.

This commit should probably be looked over by someone else, as I am not
really qualified to be writing specifications and I am ultimately
writing this specification for the RTL, not really the other way around.